requirements.txt
python-dotenv
watchfiles

# API
fastapi
uvicorn
python-multipart
httpx

# Vector db
numpy
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence_transformers
pymilvus>=2.4.2
pymilvus[milvus_lite]

# elastic
elasticsearch>=8.0.0,<9.0.0
aiohttp